+++ workflow/scripts/run_polegon_chunk.py
@@ -27,11 +27,8 @@ if use_polegon:
     params = yaml.safe_load(open(snakemake.input.params))
     params = params.pop("polegon")
     seed = params.pop("seed") + int(snakemake.wildcards.rep)
-
-    # This will match what polegon_master does. 
-    # Why the factor of 2 relative to SINGER? 
-    # Doesn't it cancel during rescaling, regardless?
-    # params["Ne"] *= 2  
+    # FIXME: Ne is lower by factor of two relative to `polegon_master`.
+    # This shouldn't matter, it cancels anyways, but look into it.
 
     # POLEGON expects inputs named slightly differently than SINGER output
     prefix = snakemake.input.muts.replace("_muts_", "_").removesuffix(".txt")
@@ -41,13 +38,14 @@ if use_polegon:
 
     # Adjust branch spans to reflect missing data, and absorb mutation rate into span.
     # This is necessary because POLEGON doesn't use the "mutation-adjusted spans"
-    # from the mutation map during rescaling. Manually editing the branch spans
-    # fixes this.
+    # from the mutation map during rescaling. Manually converting the branch spans
+    # to mutational units fixes this.
     with open(snakemake.log.out, "w") as log:
        log.write(
            "f{tag()} Adjusting branch spans input ({prefix}_branches.txt) "
            "to reflect mutational rate map and setting mutation rate to unity\n"
        )
+    params["m"] = 1.0
     mutation_map = params.pop("mutation_map")
     adjusted_mu = np.loadtxt(mutation_map, ndmin=2)
     assert np.all(adjusted_mu[:-1, 1] == adjusted_mu[1:, 0])
@@ -59,7 +57,6 @@ if use_polegon:
     for i in range(2):
        branches[:, i] = adjusted_mu.get_cumulative_mass(branches[:, i])
     np.savetxt(f"{prefix}_branches.txt", branches)
-    params["m"] = 1.0
 
     invocation = [
         f"{snakemake.params.polegon_binary}",